[ADD] lecturer, lecture, topic: add lecture, lecutrer and topic models and their views to the nl_usm and nl_ums_master_data modules. The lecturer model is inheritd form the hr.employee module and have been customized.
