+++ addons/nl_ums/__manifest__.py
@@ -25,11 +25,9 @@
     'data': [
         # 'security/ir.model.access.csv',
         'views/templates.xml',
-        'views/teacher.xml',
-
+        'views/lecturer_views.xml',
         'views/classes_views.xml',
         'views/lecture_views.xml',
-
         'views/students_views.xml',
         'views/classes_views.xml',
