+++ addons/nl_ums/models/__init__.py
@@ -3,5 +3,4 @@
 from . import lecture
 from . import classes
 from . import students
-from . import teacher
-
+from . import lecturer
