+++ addons/nl_ums_master_data/models/__init__.py
@@ -10,4 +10,3 @@ from . import building
 from . import floor
 from . import class_room
 from . import year
-
