+++ addons/nl_ums_master_data/models/semister.py
@@ -1,13 +1,12 @@
 from odoo import models, fields, api
 
+
 class Semister(models.Model):
-	
+
     _name = 'ums.semister'
     _description = 'Semister'
 
     # year_id = fields.Many2one('ums.year', 'semister_ids', string="Year")
 
-    name = fields.Selection([('1', '1'),('2', '2'),('3', '3'),('4', '4')], string="Name")
-    semister_type = fields.Selection([('spring','Spring'),('fall','Fall')], string='Type')
-    
-  
+    name = fields.Selection([('1', '1'), ('2', '2'), ('3', '3'), ('4', '4')], string="Name")
+    semister_type = fields.Selection([('spring', 'Spring'), ('fall', 'Fall')], string='Type')
